fix: Increase child process `maxBuffer` size from 1MB to 10MB
Some commands like `cat` can output significant amounts of data on stdout. This
is a hotfix to deal with this issue. I don't see an issue with increasing the
`maxBuffer` for all commands except for more memory consumption (assuming
Node.js allocates all of it ahead of spawning the process).